Extract pipes
