+++ lib/syntax/pipeline.js
@@ -24,57 +24,12 @@ var LexPipe = (function (_super) {
     return LexPipe;
 }(frames_1.Frame));
 exports.LexPipe = LexPipe;
-var EvalPipe = (function (_super) {
-    __extends(EvalPipe, _super);
-    function EvalPipe(out, meta) {
-        if (meta === void 0) { meta = frames_1.Void; }
-        var _this = this;
-        meta[frames_1.Frame.kOUT] = out;
-        _this = _super.call(this, meta) || this;
-        return _this;
-    }
-    return EvalPipe;
-}(frames_1.Frame));
-exports.EvalPipe = EvalPipe;
-var ParsePipe = (function (_super) {
-    __extends(ParsePipe, _super);
-    function ParsePipe(out, meta) {
-        if (meta === void 0) { meta = frames_1.Void; }
-        var _this = this;
-        meta[frames_1.Frame.kOUT] = out;
-        _this = _super.call(this, meta) || this;
-        return _this;
-    }
-    return ParsePipe;
-}(frames_1.Frame));
-exports.ParsePipe = ParsePipe;
-var piper = function (input, context) {
-    if (context === void 0) { context = frames_1.Void; }
-    var result = new frames_1.FrameArray([], context);
-    var evaluator = new EvalPipe(result);
-    var parser = new ParsePipe(evaluator);
-    var lexer = new LexPipe(parser);
-    var status = lexer.lex_string(input);
-    if (status !== lexer) {
-        console.error("\n* pipe returned " + status);
-    }
-    return result;
-};
 exports.framify = function (input, context) {
     if (context === void 0) { context = frames_1.Void; }
     var env = new frames_1.Frame(context);
     var codify = new frames_1.FrameLazy([]);
     var expr = pipe(input, codify);
     return expr.call(env);
-};
-exports.framify_new = function (input, context) {
-    if (context === void 0) { context = frames_1.Void; }
-    var result = new frames_1.FrameArray([], context);
-    var parser = new ParsePipe(result);
-    var status = pipe(input, parser);
-    console.error("\n* framify_new.pipe returned " + status);
-    var expr = result.at(0);
-    return expr.call(result);
 };
 var pipe = function (input, out) {
     var output = new frames_1.FrameArray([]);
